Extract version gauge and functions to separated file
